+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 0.2.0 (2019-07-16)
 
 - FEAT: change format of remote URL so that it allows setting different endpoints
+- FEAT: replace `print` with `sys.stderr.write` so that the text shows up when using git
+- FEAT: add ec2 catalog support (factored out some common code)
 
 
 Version 0.1.0 (2019-07-12)